+++ src/org/mtransit/parser/mt/data/MSpec.java
@@ -17,10 +17,10 @@ public class MSpec {
 	public List<MTripStop> tripStops;
 	public List<MServiceDate> serviceDates;
 	public Map<Integer, List<MSchedule>> stopSchedules;
-	public Map<Integer, List<MFrequency>> routeFrequencies;
+	public Map<Long, List<MFrequency>> routeFrequencies;
 
 	public MSpec(List<MStop> stops, List<MRoute> routes, List<MTrip> trips, List<MTripStop> tripStops, List<MServiceDate> serviceDates,
-			Map<Integer, List<MSchedule>> routeSchedules, Map<Integer, List<MSchedule>> stopSchedules, Map<Integer, List<MFrequency>> routeFrequencies) {
+			Map<Integer, List<MSchedule>> routeSchedules, Map<Integer, List<MSchedule>> stopSchedules, Map<Long, List<MFrequency>> routeFrequencies) {
 		this.stops = stops;
 		this.routes = routes;
 		this.trips = trips;
